requirements.txt
python-binance
websockets
smtplib
email-validator
